FFHPackageManager for Template Demo
