+++ Assets/FarFromHere/FFH Package Manager/Editor/FFHPackagesData.cs
@@ -1,41 +1,34 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-[System.Serializable]
-public class PackageActive
+namespace FFH.PackageManager
 {
-    public bool InstalledPackages { get; set; }
-    public bool EmbededPackages { get; set; }
-    public bool GUIState { get; set; }
-    public string Name;
-    public string Address;
-    public string FolderGroupLabel;
-}
-
-[CreateAssetMenu(fileName = "FFH Packages", menuName = "FFH/PackageData")]
-public class FFHPackagesData : ScriptableObject
-{
-    [SerializeField] private bool showAtStart;
-    public bool ShowAtStart
+    [System.Serializable]
+    public class PackageActive
     {
-        get => showAtStart;
-        set
-        {
-            if (showAtStart != value)
-            {
-                showAtStart = value;
-                OnShowAtStartChanged?.Invoke(showAtStart);
-            }
-        }
+        public bool InstalledPackages { get; set; }
+
+
+        public string Name;
+        public string Address;
+        public string FolderGroupLabel;
+        [HideInInspector]
+        public bool GUIState;
+        [HideInInspector]
+        public bool EmbededPackages;
     }
 
-    public event System.Action<bool> OnShowAtStartChanged;
+    [CreateAssetMenu(fileName = "FFH Packages", menuName = "FFH/PackageData")]
+    public class FFHPackagesData : ScriptableObject
+    {
+        public bool ShowAtStart;
 
-    public string PackageListDefineSymbols;
-    public bool AddedSymbols { get; set; }
+        public string PackageListDefineSymbols;
+        public bool AddedSymbols { get; set; }
 
-    public PackageActive[] Packages;
+        public PackageActive[] Packages;
 
-    public PackageActive[] ResourcesPackages;
-    public List<string> PackagesNames { get; set; } = new List<string>();
+        public PackageActive[] ResourcesPackages;
+        public List<string> PackagesNames { get; set; } = new List<string>();
+    }
 }
