+++ Assets/FarFromHere/FFH Package Manager/Editor/TemplateHDRPEnvironmentPackages/FFHPackageImporterWindow.cs
@@ -1,0 +1,10 @@
+using UnityEditor;
+
+namespace FFH.PackageManager
+{
+    public class FFHPackageImporterWindow : EditorWindow
+    {
+        public static FFHPackagesData packageData;
+        public static bool allPackagesInstalled;
+    }
+}
