+++ Assets/FarFromHere/FFH Package Manager/Editor/TemplateHDRPEnvironmentPackages/FFHPackageImporterWindow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: effc46e1d7b84ac459ac854a3936292f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 1c23823aeff990742840fed6d461615a, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
