+++ Assets/FarFromHere/FFH Package Manager/Editor/TemplateHDRPEnvironmentPackages/FFHPackageManagerUtilities.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f688a0808c246f347a45024acf2f52bc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 1c23823aeff990742840fed6d461615a, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
