+++ Assets/FarFromHere/FFH Package Manager/Editor/TemplateHDRPEnvironmentPackages/SceneReloader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e537f5e456c737b4b95099c28308e5a0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 1c23823aeff990742840fed6d461615a, type: 3}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
